+++ tango/io/selector/SelectorException.d
@@ -26,7 +26,7 @@ public class SelectorException: Exception
      * line     = line number of the source file where the exception was
      *            thrown; you would normally use __LINE__ for this parameter.
      */
-    public this(string msg, string file, uint line)
+    public this(string msg, string file, size_t line)
     {
         super(msg, file, line);
     }
@@ -48,7 +48,7 @@ public class UnregisteredConduitException: SelectorException
      * line     = line number of the source file where the exception was
      *            thrown; you would normally use __LINE__ for this parameter.
      */
-    public this(string file, uint line)
+    public this(string file, size_t line)
     {
         super("The conduit is not registered to the selector", file, line);
     }
@@ -69,7 +69,7 @@ public class RegisteredConduitException: SelectorException
      * line     = line number of the source file where the exception was
      *            thrown; you would normally use __LINE__ for this parameter.
      */
-    public this(string file, uint line)
+    public this(string file, size_t line)
     {
         super("The conduit is already registered to the selector", file, line);
     }
@@ -90,7 +90,7 @@ public class InterruptedSystemCallException: SelectorException
      * line     = line number of the source file where the exception was
      *            thrown; you would normally use __LINE__ for this parameter.
      */
-    public this(string file, uint line)
+    public this(string file, size_t line)
     {
         super("A system call was interrupted by a signal", file, line);
     }
@@ -110,7 +110,7 @@ public class OutOfMemoryException: SelectorException
      * line     = line number of the source file where the exception was
      *            thrown; you would normally use __LINE__ for this parameter.
      */
-    public this(string file, uint line)
+    public this(string file, size_t line)
     {
         super("Out of memory", file, line);
     }
